requirements.txt
langchain
langchain-google-genai
pydantic
wikipedia
python-dotenv
langchain-community